Removed more stuff
Removed all of the different lognormal analysis methods, it's still working as it should be though.
